dom/inputmethod/mochitest/test_two_inputs.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1057898
https://bugzilla.mozilla.org/show_bug.cgi?id=952741
-->
<head>
  <title>Test switching between two inputs</title>
  <script type="application/javascript;version=1.7" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.7" src="inputmethod_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1057898">Mozilla Bug 1057898</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=952741">Mozilla Bug 952741</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">

inputmethod_setup(function() {
  runTest();
});

let appFrameScript = function appFrameScript() {
  let input1 = content.document.body.firstElementChild;
  let input2 = content.document.body.children[1];

  let i = 1;

  input1.focus();

  addMessageListener('test:next', function() {
    i++;
    switch (i) {
      case 2:
        input2.focus();

        break;

      case 3:
        input2.blur();
        input2.focus();

        break;

      case 4:
        input2.blur();

        break;

      case 5:
        input2.focus();
        input2.blur();

        input1.focus();

        break;

      case 6:
        content.document.body.removeChild(input1);

        break;

      case 7:
        input2.focus();

        break;

      case 8:
        content.document.body.removeChild(input2);

        break;
    }
  });
};

function runTest() {
  let im = navigator.mozInputMethod;

  let i = 0;
  im.oninputcontextchange = function(evt) {
    var inputcontext = navigator.mozInputMethod.inputcontext;

    i++;
    switch (i) {
      // focus on the first input receives the first input context.
      case 1:
        ok(!!inputcontext, 'Receving the first input context');
        is(inputcontext.textAfterCursor, 'First');

        mm.sendAsyncMessage('test:next');
        break;

      // focus on the second input (implicitly blur the first input)
      // results the second input context.
      case 2:
        ok(!!inputcontext, 'Receving the second input context');
        is(inputcontext.textAfterCursor, 'Second');


        mm.sendAsyncMessage('test:next');
        break;

      // blur and re-focus on the second input results updated
      // input context for the second input.
      case 3:
        ok(!!inputcontext, 'Receving the second input context');
        is(inputcontext.textAfterCursor, 'Second');

        mm.sendAsyncMessage('test:next');
        break;

      // blur on the second input results null input context
      case 4:
        is(inputcontext, null, 'Receving null inputcontext');

        mm.sendAsyncMessage('test:next');
        break;

      // focus and blur on the second input sends no message;
      // focus on the first input receives the first input context.
      case 5:
        ok(!!inputcontext, 'Receving the first input context');
        is(inputcontext.textAfterCursor, 'First');

        mm.sendAsyncMessage('test:next');
        break;

      // remove on the first focused input results null input context
      case 6:
        is(inputcontext, null, 'Receving null inputcontext');

        mm.sendAsyncMessage('test:next');
        break;

      // input context for the second input.
      case 7:
        ok(!!inputcontext, 'Receving the second input context');
        is(inputcontext.textAfterCursor, 'Second');

        mm.sendAsyncMessage('test:next');
        break;

      // remove on the second focused input results null input context
      case 8:
        is(inputcontext, null, 'Receving null inputcontext');

        inputmethod_cleanup();
        break;

      default:
        ok(false, 'Receving extra inputcontextchange calls');
        inputmethod_cleanup();

        break;
    }
  };

  // Set current page as an input method.
  SpecialPowers.wrap(im).setActive(true);

  let iframe = document.createElement('iframe');
  iframe.src = 'data:text/html,<html><body><input value="First"><input value="Second"></body></html>';
  iframe.setAttribute('mozbrowser', true);
  document.body.appendChild(iframe);

  let mm = SpecialPowers.getBrowserFrameMessageManager(iframe);

  iframe.addEventListener('mozbrowserloadend', function() {
    mm.loadFrameScript('data:,(' + encodeURIComponent(appFrameScript.toString()) + ')();', false);
  });
}

</script>
</pre>
</body>
</html>
